login 구현 -완-

+++ back/deukgeun/src/main/java/com/dg/deukgeun/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dg.deukgeun.Entity.UserEntity;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long>{
+public interface UserRepository extends JpaRepository<UserEntity, Integer>{
     
     // 반환 타입은 Optional이며, 이는 조회 결과가 없을 경우 null 대신 Optional.empty()를 반환하여 NullPointer 예외를 방지
     Optional<UserEntity> findByEmail(String email);
@@ -14,6 +14,9 @@ public interface UserRepository extends JpaRepository<UserEntity, Long>{
     // 이메일 존재 여부를 확인하는 메서드
     boolean existsByEmail(String email);
 
+    // // password 존재 여부를 확인하는 메서드
+    // boolean existsByPassword(String password);
+    
      //email과 password의 존재 여부 확인 메서드
      public boolean existsByEmailAndPassword(String email, String password);
 }
